OpenSM/complib/cl_*map: Add get_next functions to the various maps
get_next searches for the 1st entry whose key is > the key specified.

As such, get_next provides for searches where an exact key is not known,
or the map may be changing between searches (and hence the key of a
previously fetched entry may no longer be in the map).

Signed-off-by: Todd Rimmer <todd.rimmer@qlogic.com>
Signed-off-by: Hal Rosenstock <halr@voltaire.com>