OpenSM: Handle "babbling" ports
A "babbling" port is a port which causes traps to be generated frequently.
It may directly be "this" port which generates the traps or the peer port
detecting the issue and that the SMA on switch port 0 generates the traps.
This has only currently been observed for trap 131 but will also apply
for traps 129 and 130 as well which are other urgent and similar traps.

Note that there appears to be a bug in Mellanox firmware for both Anafa-2 and
Tavor at a minimum which causes the max trap rate not to be adhered to
and relief for this does not appear to be in short term sight.

Policy
When a bablbing port is detected, OpenSM will disable the port or its
peer switch port (depending on which trap) which should terminate the
trap storm.

Detection
250 consecutive traps of this type will be used as the (initial)
threshold. The reason for this is so as to not prematurely detect this
and disable a port.

Recovery
Admin would reenable port when OK again. (This usually involves
rebooting the node causing the trap to be indicated.)

Signed-off-by: Hal Rosenstock <halr@voltaire.com>