PerfMgr: Clear redirection on send timeout
Send timeout may mean that the redirection information on the responder
has changed without the knowledge of the PerfMgr so clear this so that
the PerfMgr goes through the node lookup and redirection again. This may
be unneeded in some cases but there is no other good way to determine a
local LID change on the redirector side.

Signed-off-by: Hal Rosenstock <halr@voltaire.com>