opensm: Bug in coding trying to set vl_arb_high_limit
Hi Sasha

When QoS setup is done the code was trying to send updates of
vl_arb_high_limit by req_set of PORT_INFO with the new data.
However, at that stage the SM still did not assign LIDs to the ports.
So the sent PortInfo.base_lid was still zero. The specification does not
allow for such LIDs (they are considered ilegal).

the patch below fixes this by storing the calculated value and later
using it in link and lid managers.

Eitan

Signed-off-by:  Eitan Zahavi <eitan@mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>