opensm: remove reassign_lfts configuration parameter
This removes actually useless subn.opt.reassign_lfts parameter.
Its value is used only for initial setup of ignore_existing_lfts flag.
But later this flag becomes unconditionally TRUE if at least one switch
is found in the fabric. If not (and fabric is just back to back
connected CAs) there is no routing and this flag is not used. In any
case initial value does not matter.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>