opensm: remove lft setup hack
This removes the hack, where OpenSM's lfts were is updated by ucast_mgr
and not only from the network. Once it was needed for dumping fucntions,
which use the data from lft, but now the dumping is moved to the end of
the sweep, when all lfts are up to date.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>