opensm: duplicated GUID/moved port detection improvements
When during discovery phase of the OpenSM heavy sweep some already
discovered port is reconnected to different location so that OpenSM
will find it again this will be recognized as duplicated GUID
(different nodes with same guid) and OpenSM will report fatal error
and exit immediately.

This patch tries to do the better work to isolate duplicated GUID
case from regular port moving as described above. This will resend
NodeInfo query to the old location. This doesn't resolve all possible
reconnections, for instance:

A: sm -> .. -> s1:N -> h1    B: sm -> .. -> s1:N -> h1
               s1:M -> s2                   s1:M -> s1

If such change was done after switch s2 was discovered node h1 will
be detectable at both paths original (via s1) and new one (via s2,
which is s1 now).

In order to solve this in the case when targeted node is switch we
will query NodeInfo and match remote port and not targeted node itself.
Unfortunately it cannot be used with CA because it doesn't forward MADs,
so when check fails we will request for new discovery.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>