opensm: resweep instead of exit when duplicated guid suspected
Anyway OpenSM will request resweep when there are suspected nodes
with duplicate GUID on the subnet. And because we cannot be 100% sure
that detected GUIDs duplication is not some corner case of port moving
I prefer to not exit. Endless (re)discovery and syslog messages should
be good indication if it is indeed this case.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>