opensm: set scope according to configuration for ipoib multicast groups
Hi Sasha,

opensm creates two multicast groups for ipoib.  The group corresponding to
255.255.255.255 results in an MGID with the scope as specified by the configuration.
The one MGID for 0.0.0.1 ends up with local scope irrespective of the configuration.
If ipoib is configured with global scope, then its join of the latter group never
succeeds.  The patch below sets the scope for both groups according to the
configuration.

Signed-off-by: Rolf Manderscheid <rvm@obsidianresearch.com>

----
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>