opensm: support multiple routers in a subnet
Hi Sasha,

If a path record query is made for an off-subnet DGID, the SA needs to
return a path record where the DLID points to the router port that
handles the DGID prefix.  In the case of a subnet with only one
router, the SA could just pick "the router", and that's exactly what
the ROUTER_EXP code did.  However, ROUTER_EXP did not look beyond the
first available router.

When additional routers are added, the SA needs more information.  The
mechanism for gathering this information has not yet been specified,
so in the meantime, this patch adds a configuration file that
specifies which router ports handle which prefixes.

The patch also removes all occurrences of ROUTER_EXP ifdefs.  The
default behaviour remains unchanged with one minor exception:
hop limits are set to 0xFF for path records to multicast DGIDs if
the scope is non-local and to unicast DGIDs if off-subnet.
This used to happen for ROUTER_EXP only.

Now, the same binary can be configured at run-time to enable the
ROUTER_EXP behaviour with a generic configuration file, or to handle
multiple routers on a subnet with a more explicit configuration file.
See the man page for details.

Signed-off-by: Rolf Manderscheid <rvm@obsidianresearch.com>

--

One consequence of this patch is that people accustomed to using
ROUTER_EXP will need to specify a configuration file to get the
same behaviour.  I toyed with the idea of keeping one ROUTER_EXP ifdef
to control the default behaviour, but then we're back to having two
versions of opensm with different default behaviours, and the
counter-intuitive: empty cfg file != non-existent cfg file.  One of
the goals was to get to a single standard binary.  So, to help avoid
surprises, I actually added back one ifdef ROUTER_EXP which causes the
compilation to fail with a useful message.  This only helps those who
both build and configure their special ROUTER_EXP opensms, but I
suspect that's most.

Thanks to Hal for reviewing early versions of this patch and providing
feedback.

    Rolf
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>