osm: QoS - adding simplified syntax for policy definition
This patch adds simplified syntax for QoS definition in QoS Policy file.
Using this syntax the administrator is able to define QoS policy per
ULP and/or per Service ID and/or per partition (pkey).

Here's an example of the policy file with the new syntax, which is added
in a new section called qos-policy:

qos-ulps
     sdp, port-num 30000                   : 1 #SL for SDP when destination port is 30000
     sdp, port-num 10000-20000, 0xfffd     : 2
     sdp                                   : 0 #default SL for SDP
     srp, target-port-guid 0x1234-0x1235   : 2
     iser, port-num 0x3234-0x3235          : 4 #SL for iSER whith specific target ports
     iser                                  : 5 #default SL for iSER
     rds, port-num 25000                   : 2 #SL for RDS when destination port is 25000
     rds                                   : 0 #default SL for RDS
     ipoib, pkey 0x0001                    : 5 #SL for IPoIB on partition with pkey 0x0001
     ipoib                                 : 6 #default IPoIB partition - pkey=0x7FFF
     any, pkey 0x0ABC                      : 3
     any, pkey 0x0ABD-0x0ABF,0x0BBD-0x0BBA : 4
     any, service-id 0x6234                : 2
     any, target-port-guid 0x2234-0xF235   : 2
     default                               : 0 #default SL
end-qos-ulps

Since any section of the policy file is optional as long as basic rules
of the file are kept, the above example can serve as a complete QoS
policy file - short and clear.

I suspect that most of the administrators will use only this syntax, but
if someone wishes to manage QoS in more detailed manner, there's always
the rest of the policy file to do so.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>