opensm: Fixing broken logic in 'process world' part of LinkRecord processing
Fixing broken logic in 'process world' part of LinkRecord processing.
When both HCA's ports belong to the same subnet, OpenSM would scan
'half-world' for each port of this HCA, and then for each port it would
get the node and iterate again through all the ports of this node.
In addition to the time consumed by these unnecessary iterations, it
also caused some records to be found twice.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>