opensm: recover only for base LID values >= 0xc000
In initial stage of a subnet value of LID = 0 in PortInfo response can be
valid. Also when LID was zeroed externally OpenSM will ignore this event
(and it was the reason for regression shown in recent ibmgtsim runs). So
we will try to recover LID value only when it is >= 0xc000. This should
fix bug 246 as well and preserves original OpenSM LID "trimmer" behavior
for the case of LID = 0.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>