opensm/osm_pkey_mgr.c: setting only outbound partition enforcement on switch
Fixing wrong setting of partition enforcement bits on switch ports.
When an HCA port is configured with a certain pkey, the peer port
on the switch should turn on outbound partition enforcement bit only.
Turning on the inbound enforcement will cause the switch to drop
valid packets if the HCA is partial member.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>