opensm/osm_qos_policy.c: trivial fix in passing wrong pointer
st_lookup() returned node in the p_node pointer and replaced
the node that was intended to be inserted into the queue,
which caused infinite loop.
Besides, if st_lookup() does finds an element in the hash,
we're only interested to know that it did - don't need the
actual element.

Please apply to ofed_1_3 and master.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>