opensm/osm_ucast_ftree.c: fixing coredump in fat-tree routing
Hi Sasha,

Fat-tree routing wasn't clearing the data structure at the
beginning of the new routing recalculation, which was causing
wrong routing and sometimes coredump.

Please apply this patch to master and ofed_1_3.
If it's possible, please apply to ofed_1_3 ASAP - we're doing
RC2 today, and it's important to have this fix included.

-- Yevgeny

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>