opensm: osm_subnet.c log and print the path of the cached option file
Sasha,

A trivial patch.  During development, we (at LLNL) sometimes use different
options/configurations.  This provides a way to know which one is active.

 From d60408b4dc1cb0c917e2eb33d6a3f62ac6bb9b5c Mon Sep 17 00:00:00 2001
From: Tim Meier <meier3@llnl.gov>
Date: Thu, 24 Jan 2008 11:51:08 -0800
Subject: [PATCH] opensm: osm_subnet.c  log and print the path of the cached option file

Logged (syslog and print) the path to the option file that is cached and parsed.
This is helpful when something other than the default path is used.

Signed-off-by: Tim Meier <meier3@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>