opensm: in UP/DOWN algo compare GUID values in host byte order
In Up/Down direction is determined by comparing GUID values when
destination to root nodes is equal. This comparison was done in network
byte order - I found this as very confused when tried to use this
feature. Another bad things with it is that UP/DOWN algorithm in OpenSM
may work differently on big- and little- endian machines. Fixing this.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>