opensm: updn/connect_roots: preserve connectivity to root nodes
Option '--connect_roots' was added in order to preserve connectivity
between root nodes when Up/Down routing algorithm is used (by default
Up/Down leaves root nodes not connected).

The problems with this option could appear with some topologies when
some root's (root A) neighbor switch has valid up/down path (which
could not be shortest path due to up/down restriction) to another root
node (root B) in a fabric which cross this root A node and at the same
time root A has shortest path (preserved with --connect_roots option)
to the root B which crosses its neighbor.

In this case we will have single routing loop which is not good.
(actually it causes OpenSM to crash during multicast spanning trees
building).

Solution is to preserve to root nodes shortest paths (on when
--connect_root is specified) for all switches, not only other roots as
it is today.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>