opensm/QoS: setting SL in the IPoIB MCast groups
Hi Sasha,

This is a reworked patch that sets SL in the IPoIB mcast groups
Note that I'm using functions here! :)

Added mcast mlid to the partition structure.
This mlid is used later by the QoS manager for quick access
to the mcast group.

The patch is for master only.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>