OpenSM: Add a Performance Manager HOWTO to the docs and the dist
There seems to be a lot of questions on the list about how to gather port
counters.  The Performance Manager included in OpenSM, v3.1.X (OFED 1.3) can be
used to collect these counters in one place.  I decided to write a little HOWTO
to help people to set it up.

Patch is attached,
Ira

Signed-off-by: Ira K. Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>