opensm: preserve base lid routes
Basically this addresses the problem described by Al Chu in:

http://lists.openfabrics.org/pipermail/general/2008-April/049132.html

When base lid paths become completely disbalanced on a fabrics with
lmc > 0.

One feedback was from Yiftah Shahar:

"I think that our requirements should be that even when you are working
with LMC>0 then the base LID routing should not be affected.
One way to achieve this goal is to first run the base-LID routing (so
all base LID improvement will be also in LMC>0) and then start with the
other LIDs as round-robbing starting from the base-lid-port + 1
according current routing algorithm rules (keeping min-hop, up/down...)."

We had some discussion with Al and Yiftah about this and considered that
in addition to "pure" base lid paths preservation (which is good thing by
itself) proposed method solves original lid disbalancing problem as well.

This patch is implementation of the idea above.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>