opensm: reroute console option
Hey Sasha,

With all of these file-input possibilities for OpenSM routing, I don't
think there is a way to tell OpenSM to reroute if (for example) the file
input has changed but the network/fabric has not.  This is a patch to
add a command to the console to tell the console to reroute but not do a
heavy sweep.

Longer term, perhaps there could be a plugin/thread/something to watch
the routing input files and see if they change, then set the
"force_reroute" flag as needed.  But I think this console option should
suffice for now.

Signed-off-by: Albert L. Chu <chu11@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>