opensm: remove -rpath from LDFLAGS
Fedora is very picky that things pass rpmlint before they are allowed
into the distribution.  They only allow things that rpmlint flags as an
error in the case of obvious necessity.  One of the things that came up
in the review process is that opensm still uses rpath settings in its
various Makefile.am files.  The use of rpath at all is considered to be
an error and is currently blocking things.  I went in to the Makefile.am
files and attempted to clean them up.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>