opensm: Add a Node Description check on light sweep.
A Node Description check on light sweep will ensure that the ND has been found
for each node.  This case covers the condition where a ND message is
dropped/lost for some reason and OpenSM is left with a valid configured node
which is not named correctly.

This is not the same as a node which has changed it's Node Descriptioin.  In
this case the node needs to send a trap.

Signed-off-by: Ira Weiny <weiny2@llnl.gov>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>