opensm/osm_config.h: generated OSM_DEBUG macro
In accordance with requested mode ./configure will generate OSM_DEBUG
macro in files config.h and osm_config.h. This will be defined as '1'
when debug mode was enabled and undefined otherwise. It can be used by
third parties (such as ibutils and plugins) to know about OpenSM build
mode.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>