opensm/Unicast Routing Cache: add osm_ucast_cache.{c,h} files
Implementation of the Unicast Routing Cache.

Unicast Manager stores all the links and ports that went down in
order to avoid routing recalculation whenever is possible.
There is a map of the switches that went down or have one or more
ports that went down. For ports the only thing that is cached is
connectivity (remote side lid and type). If the whole switch went
down/disconnected, ucast cache also stores all its unicast routing
information (lft, max_lid, lid matrices). This information will be
restored to the switch when it will be back online.

When new link/node is spotted during the discovery process, ucast
cache checks whether it was previously cached or is it a brand new
link/node and decides whether or not new routing calculation needed.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>