opensm/Unicast Routing Cache: integrate cache into opensm
Integrating unicast cache into the discovery and ucast manager.
Unicast cache is integrated as part of osm_ucast_mgt_t object.
The cache is marked valid after the first routing engine
execution.

During the discovery or in the drop mgr, whenever some node/port
goes down/disconnected, it is cached in the unicast cache.
Whenever new node or port were spotted, unicast cache checks
whether this is a new node/port or is it something that was
previously cached.

As long as fabric changes do not require full routing recalculation,
unicast cache will prevent osm_ucast_mgr_process() execution (which
also prevents routing engine execution).
When the topology change is not something that can be tolerated
by the unicast cache, cache is invalidated, all its data structures
are freed, new routing is calculated by the routing engine, and the
cache is marked as valid again.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>