opensm: sweep on SIGCONT
When OpenSM is suspended by SIGSTOP it can miss subnet changes. Example
of known scenario is when another Stanby SM becomes master due to
unresponsive suspended master SM, when suspended SM continues to run
(after SIGCONT) we have two master SMs in a subnet.

To fix this we will schedule heavy sweep when SIGCONT is received.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>