opensm: fix iser service-id used for SL assignment
RFC3720 says:

The well-known user TCP port number for iSCSI connections assigned by IANA is 3260
and this is the default iSCSI port. Implementations needing a system TCP port number
may use port 860, the port assigned by IANA as the iSCSI system port; however in
order to use port 860, it MUST be explicitly specified - implementations MUST NOT
default to use of port 860, as 3260 is the only allowed default.

Hence the SID used by iser is 0x0000000001060CBC and not 0x000000000106035C

Signed-off-by: Or Gerlitz  <ogerlitz@voltaire.com>
Signed-off-by: Eli Dorfman <elid@voltaire.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>