opensm/osm_lid_mgr.c: ignore and overwrite guid2lid (windows)
Hi Sasha,

When Windows is crashing with BSOD, it might corrupt files that were
previously opened for writing, even if the files are closed. As a result,
we might see corrupted guid2lid file, and OpenSM will exit on such error.
This patch makes SM ignore (and later overwrite) corrupted guid2lid files.

The patch has already been accepted into ofw.

I'm posting it to openib too, so that when some day WinSM will be
synchronized with OpenSM, this fix won't be lost.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>