opensm/osm_port_info_rcv.c: Remove SM from sm_guid_tbl when IsSM is not
Sasha,

The following patch removes the SM from the sm_guid_table when IsSM is
not present.
v2 of this fixes the memory leak you pointed out in the original
version. Compile tested only as I don't have an environment to recreate
this anymore.

-- Hal

opensm/osm_port_info_rcv.c: Remove SM from sm_guid_tbl when IsSM is
not present in PortInfo:CapabilityMask

SM should be removed from the sm_guid_tbl subsequent to a trap 144
indicating the capability mask changed (and the new capabilities
no longer include IsSM).

As a result of this, move clearing of SM state to be conditionalized on
IsSM present rather than regardless of whether IsSM is set.

Prior to this patch, the OpenSM log is spammed with error messages on
SubnGets of SMInfo attribute.

Signed-off-by: Hal Rosenstock <hal.rosenstock@gmail.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>