mesh analysis - lash preparation
Sasha,

Here is a revised mesh patch #9 that incorporates changes based on your
comments.

This patch makes some minor cleanups in osm_ucast_lash.c in preparation for
next steps.

The main change is to minimize the occurrences of phys_connections. These
changes are mainly cosmetic and independent of the mash analysis code but
will make patch 10 easier to understand.

Regards,

Bob Pearson

Signed-off-by: Bob Pearson <rpearson@systemfabricworks.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>