mesh analysis - integrate into lash core
Sasha,

Here is a revised mesh patch #10 that incorporates changes based on your
comments.

This patch
      - hooks mesh code into lash
      - replaces sw->phys_connections by the equivalent switch->node->links
      - replaces sw->num_connections by the equivalent
switch->node->num_links
      - replaces sw->virtual_physical_port_table by
switch->node->links[]->ports

When the do_mesh_analysis flag is not set there is no change to the function
except To replace the variables with variables in node that have the same
size. In this Case the port table in link_t will always have just one port.

When the do_mesh_analysis flag is set multiple physical links will collapse
to a Single logical link with a port list with more than one element.

      - rewrote connect switches to use variables in node
      - in log Lane requirements (%d) exceed available lanes (%d)
        Arguments were reversed, fixed
      - compute physical egress port in routine get_next_port
        Which will use round robin if there are more than one
        Physical links between switches

Regards,

Bob Pearson

Signed-off-by: Bob Pearson <rpearson@systemfabricworks.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>