Corrected incoherency in __osm_ftree_fabric_route_to_non_cns comments
It seems to be that there is an error in the comment of __osm_ftree_fabric_route_to_non_cns comments.
It said the function was to be called with TRUE,FALSE parameters when it was called with TRUE,TRUE.
I'm just discovering the Ftree routing code so I may have misunderstood something.

Anyway if I'm right, here's the patch!

Signed-off-by: Nicolas Morey-Chaisemartin <nicolas.morey-chaisemartin@ext.bull.net>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>