opensm: Implement weighted routing
Implement a weighted routing scheme for fine tuning the lid matrix for
routing engines that use the lid matrix.  An optional file containing
a switch_guid port and weighing factor combination per line can be
supplied to override a default hop weight factor of 1 for each switch
output port in computing the lid matrix.  This allows one to alter the
min hop paths for things like routes to I/O.

Signed-off-by: Dale Purdy <purdy@sgi.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>