opensm/osm_link_mgr.c initialize SMSL
The SMSL is currently not initialized. Whatever is written in PortInfo
at power up is kept.

The lash routing algorithm makes active use of SL to avoid deadlock in
the fabric. The endnodes needs to be instructed to use the SL calculated
for the endnode-to-SMnode path for SA traffic. If not, the SA traffic
can cause deadlock.

This change makes sure that the SMSL is properly initialized on all
ports. If flash routing is used, the SMSL is set up based on routing
calculations. For other routing algorithms the default SMSL is used.

Signed-off-by: Line Holen <line.holen@sun.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>