opensm/osm_subnet.c: fixing compiler warnings
Hi Sasha,

There are couple of meaningless CL_ASSERT() lines in the
parsing functions that are checking the parsed value.
The compiler says that "comparison is always true due
to limited range of data type". So either the 'val' type
should be more general when parsing it, or the CL_ASSERT()
lines are not needed. The following patch removes them.

Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>