opensm/osm_ucat_ftree.c Enhance min hops counters usage
This patch enhances the use of the min hop table done in the Fat-Tree algorithm.
Before this patch, the algorithm was using the osm_sw hops table to store the minhop values toward any lid (Switch or not).
As this table is allocated as we need it, it required a lot of malloc calls and quite some time to set the hops values on remote ports.

This patch corrects this behaviour:
-The osm_sw hops table is only used for switch lid
-ftree_sw_t struct now has its own hop table (only 1 dimensionnal as we don't need to know which port is used) to store its minhop value

Signed-off-by: Nicolas Morey-Chaisemartin <nicolas.morey-chaisemartin@ext.bull.net>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>