opensm/osm_ucast_ftree.c: Fixed issue with reverse_hops introduced by commit 1d7dd18b531c1d6370f80cc7303493d6f3e3e777.
When using reverse hop with nodes on root switches, origin switch is explored multiple times with wrong hop numbers creating
loops in the network.
This is fixed by checking before going up (even on the master path) that the remote switch is either not
configured or with a longer path (second test was missing).

Signed-off-by: Nicolas Morey-Chaisemartin <nicolas.morey-chaisemartin@ext.bull.net>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>