opensm/osm_mesh.c: Improve VL utilization
opensm/osm_mesh.c: Improve VL utilization

Add steps to the algorithm to further condition the netlist of switches
that is handed to lash to analyze.

(1)  The previous version sorted out all the links out of switches into
dimension order’in the order the dimensions were found which is driven by
the arbitrary choice of links out of the seed’switch. This takes the
additional step of waiting until after the lengths along each dimension
are measured then reordering the links out of each switch so that dimensions
with the largest lengths are ordered first. i.e. if the mesh analysis discovers
that the mesh is a 5x8x6 torus then the coordinates are reordered so that based
on the order of links on each switch the mesh is an 8x6x5 torus.

(2) The previous version left the order of switches in the switch list the
same as when they were discovered by the SMPs which is some sort of depth
first tree walk modified by the random return of MADs if multiples are
outstanding. This takes the additional step of reordering the switches
so that they are presented in odometer’order again with the dimension
with the largest lengths changing fastest.

With this algorithm, we always see what we believe is the optimal result
for each size of toroidal mesh.

As a footnote, when we later studied the effect of broken or imperfect meshes,
we found that the results are dependent on the choice of origin for the mesh
which clearly doesn't matter for a perfect torus with nodes sorted as described
here. Thus there is some additional work required to try to understand how to
select the optimal origin for meshes with defects. We saw that the number of
VLs required was higher when the origin was near a defect so choosing one as
far as possible from the defects seems to be the right idea. This still needs
to be done since it will remove some fairly rare defect cases which now require 9 VLs to route.

Signed-off-by: Robert Pearson <rpearson@systemfabricworks.com>
Signed-off-by: Hal Rosenstock <hal.rosenstock@gmail.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>