opensm/osm_ucast_ftree.cd: Added support for same level links
This patch adds support for horizontal (sibling) links in the Fat-Tree algorithm.
Horizontal links are handled as both upward and downward. They are treated as any other through the exploration but they are only considered last to ensure that for 2 paths of the same lengths the one without horizontal links is preferred.

Also trying horizontal links after other links ensures that the horizontal path will be as far as possible in the created route (in the ftree exploration sense) thus reducing (removing?) the risk of credit loops.
Several tests have been run with Sun QNEM switches and results are good.

This patch may require some enhancements later (horizontal links in higher levels have not been fully validated) but it does not change the behaviour on topologies without horizontal links. Thus it should be safe to apply.

Acked-by: Line Holen <Line.Holen@Sun.COM>
Signed-off-by: Nicolas Morey-Chaisemartin <nicolas.morey-chaisemartin@ext.bull.net>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>