opensm/osm_mesh.c: Reorder switches for lash
The goal of this patch is to change the order of the switches in the array kept
in the lash context from the original order to one in which the switches are
presented in 'odometer order'.

When the main routine in lash is called the switches are in an order that is
likely based on the order that the switches were originally visited by SM
topology discovery which is some sort of tree walk. All of the analysis up to
this point is independent of the actual order of the switches, but lash will
use that order to enumerate the paths in the fabric and add them to the VL bins.

Odometer order means that the switches are labelled s[X0, ..., Xn-1] and
ordered s[0, ..., 0], s[0, ..., 1], s[0, ..., Ln-1], s[0, .. 1, 0] etc.
The dimensions are also reordered so that the dimension changing the fastest
has the largest length, i.e. Ln >= Ln-1 >= ... >= L1. [All this is modulo
possible end to end reversal but the basic idea is that the longest axis
changes fastest.]

TO INVESTIGATE: Rather than using an additional switches array in
sort_switches whether it can be done in place using p_lash->switches.

Signed-off-by: Robert Pearson <rpearson@systemfabricworks.com>
Signed-off-by: Hal Rosenstock <hal.rosenstock@gmail.com>
Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>